pyproject.toml
[project]
name = "myquickfoldertagger"
version = "0.1.0"
description = "A locally-run, cross-platform CLI tool that reads a target codebase's README.md and uses an LLM to assign a 1-3 word tag describing what the codebase does. Results are stored in a configurable CSV file."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32.5",
    "openai>=2.28.0",
    "anthropic>=0.84.0",
    "google-genai>=1.67.0",
]
